# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: comparisons
    description: Vendor quotation comparison operations
  - name: vendors
    description: Vendor management and quotation insights
  - name: billing
    description: Farrez Pro subscription operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /comparisons:
    get:
      operationId: listComparisons
      tags: [comparisons]
      summary: List saved comparisons
      parameters:
        - {name: search, in: query, schema: {type: string, maxLength: 100}}
        - {name: sort, in: query, schema: {type: string, enum: [createdAt, vendorName]}}
        - {name: order, in: query, schema: {type: string, enum: [asc, desc], default: desc}}
      responses:
        "200":
          description: Saved comparisons
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Comparison"}}
    post:
      operationId: createComparison
      tags: [comparisons]
      summary: Compare and save vendor quotations
      description: Calculates totals, recommends the lowest-cost vendor with delivery speed as the tie-breaker, and stores the comparison in MongoDB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComparisonInput"
      responses:
        "201":
          description: Comparison calculated and saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "400":
          description: Invalid comparison input
  /comparisons/{id}:
    get:
      operationId: getComparison
      tags: [comparisons]
      summary: Get a saved comparison
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Saved comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "400":
          description: Invalid comparison identifier
        "404":
          description: Comparison not found
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      parameters:
        - {name: search, in: query, schema: {type: string, maxLength: 100}}
        - {name: status, in: query, schema: {type: string, enum: [active, inactive]}}
        - {name: sort, in: query, schema: {type: string, enum: [name, createdAt, updatedAt]}}
        - {name: order, in: query, schema: {type: string, enum: [asc, desc], default: asc}}
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Vendor"}}
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VendorInput"}
      responses:
        "201":
          description: Created vendor
          content:
            application/json: {schema: {$ref: "#/components/schemas/Vendor"}}
        "400": {description: Invalid vendor}
        "409": {description: Vendor already exists}
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Vendor detail and insights
          content:
            application/json: {schema: {$ref: "#/components/schemas/VendorDetail"}}
        "400": {description: Invalid vendor identifier}
        "404": {description: Vendor not found}
    patch:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VendorUpdate"}
      responses:
        "200":
          description: Updated vendor
          content:
            application/json: {schema: {$ref: "#/components/schemas/Vendor"}}
        "400": {description: Invalid input}
        "404": {description: Vendor not found}
        "409": {description: Vendor already exists}
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Archive a vendor
      description: Marks the vendor inactive without deleting historical comparison data.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Deletion result
          content:
            application/json: {schema: {$ref: "#/components/schemas/DeleteResult"}}
        "404": {description: Vendor not found}
  /billing/plan:
    get:
      operationId: getBillingPlan
      tags: [billing]
      summary: Get the active Farrez Pro subscription plan
      responses:
        "200":
          description: Active subscription plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPlan"
        "404":
          description: Farrez Pro plan not found
  /billing/checkout:
    post:
      operationId: createBillingCheckout
      tags: [billing]
      summary: Create a Stripe Checkout Session for Farrez Pro
      responses:
        "201":
          description: Checkout Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "404":
          description: Farrez Pro plan not found
  /billing/checkout-session:
    get:
      operationId: getBillingCheckoutSession
      tags: [billing]
      summary: Verify a Stripe Checkout Session after redirect
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Checkout Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionStatus"
        "400":
          description: Invalid session identifier
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    VendorQuoteInput:
      type: object
      properties:
        id:
          type: number
        vendorName:
          type: string
          minLength: 1
        quotedPrice:
          type: number
          minimum: 0
        additionalFees:
          type: number
          minimum: 0
        deliveryTime:
          type: number
          exclusiveMinimum: 0
        paymentTerms:
          type: string
          minLength: 1
        vendorId:
          type: string
      required:
        - id
        - vendorName
        - quotedPrice
        - additionalFees
        - deliveryTime
        - paymentTerms
      additionalProperties: false
    ComparisonInput:
      type: object
      properties:
        vendors:
          type: array
          minItems: 2
          maxItems: 3
          items:
            $ref: "#/components/schemas/VendorQuoteInput"
      required:
        - vendors
    Vendor:
      type: object
      required: [id, name, commercialLicense, category, status, createdAt, updatedAt]
      properties:
        id: {type: string}
        name: {type: string, minLength: 1}
        commercialLicense: {type: string, minLength: 1}
        category: {type: string, minLength: 1}
        status: {type: string, enum: [active, inactive]}
        notes: {type: ["string", "null"]}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    VendorInput:
      type: object
      required: [name, commercialLicense, category]
      properties:
        name: {type: string, minLength: 1}
        commercialLicense: {type: string, minLength: 1}
        category: {type: string, minLength: 1}
        status: {type: string, enum: [active, inactive]}
        notes: {type: string}
    VendorUpdate:
      type: object
      properties:
        name: {type: string, minLength: 1}
        commercialLicense: {type: string, minLength: 1}
        category: {type: string, minLength: 1}
        status: {type: string, enum: [active, inactive]}
        notes: {type: string}
    VendorQuoteSnapshot:
      type: object
      required: [comparisonId, date, quotedPrice, totalCost, deliveryDays, outcome]
      properties:
        comparisonId: {type: string}
        date: {type: string, format: date-time}
        quotedPrice: {type: number}
        totalCost: {type: number}
        deliveryDays: {type: number}
        outcome: {type: string, enum: [recommended, notRecommended]}
    VendorInsight:
      type: object
      required: [comparisonCount, recommendationCount, recommendationRate, averageQuotedPrice, averageTotalCost, averageDeliveryDays, latestQuote, previousQuotes]
      properties:
        comparisonCount: {type: integer}
        recommendationCount: {type: integer}
        recommendationRate: {type: number}
        averageQuotedPrice: {type: number}
        averageTotalCost: {type: number}
        averageDeliveryDays: {type: number}
        latestQuote: {type: ["object", "null"], allOf: [{$ref: "#/components/schemas/VendorQuoteSnapshot"}]}
        previousQuotes: {type: array, items: {$ref: "#/components/schemas/VendorQuoteSnapshot"}}
        pricePercentageChange: {type: ["number", "null"]}
        totalCostPercentageChange: {type: ["number", "null"]}
    VendorDetail:
      allOf:
        - {$ref: "#/components/schemas/Vendor"}
        - type: object
          required: [insights]
          properties:
            insights: {$ref: "#/components/schemas/VendorInsight"}
    DeleteResult:
      type: object
      required: [id, deleted]
      properties:
        id: {type: string}
        deleted: {type: boolean}
    VendorQuoteResult:
      allOf:
        - $ref: "#/components/schemas/VendorQuoteInput"
        - type: object
          properties:
            totalCost:
              type: number
            isRecommended:
              type: boolean
            isLowestCost:
              type: boolean
            isFastestDelivery:
              type: boolean
          required:
            - totalCost
            - isRecommended
            - isLowestCost
            - isFastestDelivery
    Comparison:
      type: object
      properties:
        id:
          type: string
        vendors:
          type: array
          minItems: 2
          maxItems: 3
          items:
            $ref: "#/components/schemas/VendorQuoteResult"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - vendors
        - createdAt
    BillingPlan:
      type: object
      properties:
        productId:
          type: string
        priceId:
          type: string
        name:
          type: string
        description:
          type: string
        unitAmount:
          type: integer
          description: Amount in Stripe's minor currency unit.
        currency:
          type: string
        interval:
          type: string
          enum: [month]
      required:
        - productId
        - priceId
        - name
        - description
        - unitAmount
        - currency
        - interval
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
          format: uri
      required:
        - url
    CheckoutSessionStatus:
      type: object
      properties:
        status:
          type: string
          enum: [open, complete, expired]
        paymentStatus:
          type: string
          enum: [paid, unpaid, no_payment_required]
        subscriptionStatus:
          type: string
          nullable: true
      required:
        - status
        - paymentStatus
        - subscriptionStatus