# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VendorStatus } from './vendorStatus';

export interface Vendor {
  id: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  commercialLicense: string;
  /** @minLength 1 */
  category: string;
  status: VendorStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}