# lib/api-zod/src/generated/types/vendorInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VendorInputStatus } from './vendorInputStatus';

export interface VendorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  commercialLicense: string;
  /** @minLength 1 */
  category: string;
  status?: VendorInputStatus;
  notes?: string;
}